requirements.txt
google-adk
mcp[cli]
httpx
python-dotenv
uvicorn
fastapi
pydantic